requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
